[CALCITE-3955] Remove the first operand of RexCall from SqlWindowTableFunction
In CALCITE-3382, we introduced TUMBLE window function to replace the
deprecated group tumble window.

But for query

```sql
select *
from table(tumble(table Shipments, descriptor(rowtime),
  INTERVAL '1' MINUTE))
```
the output plan is

```xml
LogicalProject
  LogicalTableFunctionScan(invocation=[TUMBLE($1, ...)], rowType=...)
      LogicalProject
            LogicalTableScan
```
The first operand of TUMBLE rex call is always the last
input field, but actually it represents the source table
which is the input rel node.

Removes the first operand from the RexCall because
it is useless and confusing.


Former-commit-id: f62d6b9a2f3dd877d4e2e5cbd2ff5d0d5f0efff1