[CALCITE-3936] JDBC adapter, when generating SQL, changes target of ambiguous HAVING clause with a Project on Filter on Aggregate
We achieve this with a significant change to SqlImplementor:
before you convert a RelNode to a SqlNode that represents a
SELECT, you now have to declare up-front which clauses you
are going to add. This gives SqlImplementor the information
to decide where to make the 'cut' between a query and a
sub-query, and therefore it knows what column aliases are
going to be in scope in a GROUP BY or HAVING clause.

Rename 'visitChild' method to 'visitInput';
deprecate varargs version of 'builder' method.

Close apache/calcite#2047


Former-commit-id: 8bfc3804ebcf08a9b4015be5f9b7dac92cdbe6fa