[CALCITE-3966] Trigger rules for existing RelSubset when it becomes delivered
Existing RelSubset may be first created by parent trait request, so its state
is required. But when there is a new RelNode in the same RelSet can deliver the
same traitset, we mark the subset as derived/delivered too, in which case, we
should also fire rule for the subset.

In addition, rename DERIVED to DELIVERED.

Close #1956


Former-commit-id: e7224147916d5a9bf43c322894c04b302b3a545d