[CALCITE-3980] Redis-adapter redis connection is not reused when RedisTable is created (Xu Zhang)
When creating RedisTable via ModelHandler, it relies on an instance method
RedisEnumerator#deduceRowType to extract field type map, redis connection will
be established and closed every time a table is created, which is unnecessary
and redundant, so simply make the method a static one will solve the problem
and make it more performant. CsvTable implementation works the right way.

Close #1965


Former-commit-id: 09b13cf6c61739d3ea15adf80b9dc0bc2b1ad0a0