[CALCITE-4011] Support trait propagation for EnumerableProject and EnumerableFilter (Rui Wang)
Trait propgation includes trait passthrough and trait derivation.

Trait passthrough could happen when Project satisfies the ordering
requirement that is defined by collations. Project will not satisfy the
ordering requirement when a requested collation is defined on a
non-trivial expression. Usually a RexCall is considered as non-trivial,
unless it is a CAST that perserves monotonicity,

Here is an example to demonstrate why trait cannot pass through when
collations are defined on non-trival expr:

select a, b*-1 as b
from foo
order by a, b;

which generates the logical plan:
LogicalSort
  LogicalProject
     LogicalTableScan

We cannot move the top sort down through the project. Because b*-1 will
change ordering to the opposite. The sort has to remain on top of
project for correness.

Trait derivaiton does something simlar to trait pass through, except for
one difference:
trait derivation can return parital collations that are derived from
child. For example, if [a, b, c, d] is derived from child, and if c is
defined on a non-trival expr, then [a, b] will be returned cause it
might be useful for parents.

Another example to show why trait derivation could be useful to reutrn
partial collations:

select a, b
from (
 select a, b, c*-1, d
 from foo
 order by a, b, c, d
)
order by a, b;

In this example, even though the inner project does not preserve the
total ordering for inner sort, but the outer sort only want to sort on
[a, b], thus if inner project can derive [a, b], the outer project will
not need the top sort or enforce a sort for its input.

After top-down optimization is enabled, trait propagation for EnumerableProject
and EnumerableFilter is supposed to replace ProjectSortTransposeRule and
SortProjectTransposeRule.

Close #1985


Former-commit-id: 0af3fd17a293d37125c7cca58257e5f6cbc1a76c