Refactor DDL execution code out of SqlNode sub-classes and into DdlExecutor
In the previous commit we moved DDL classes from server to
core, but in this commit we move the execution logic back to
server.

DdlExecutor is acquired via the parser factory. Not ideal, but keeps
everything working and can be improved later.

Remove duplicate package-info files, and suppress a checkstyle
complaint that the org.apache.calcite.server package in the
server module should have a package-info. (It already has one
in the core module.)


Former-commit-id: c0f102d2013a8f4d2903ec259449f5bd668eec30