[CALCITE-4015] Pass through parent collation request on subset or superset of join keys for EnumerableMergeJoin.
Collations can pass through even when required collations are subset or superset on join keys.

For subset case, collations must be extended to make sure join keys are all sorted. Collations must be the prefix of sort keys.

For superset case, collations must be either all defined on left join input, or right join input. Collations must be prefix of join keys. All collations can be pushed down to the side on which collations are fully defiend. Only collations that are defined on join keys can be pushed down to another side.


Former-commit-id: 88d18185e6177c9df587bdd23dd4049f59adc2e4