[CALCITE-4034] InnoDB adapter (neoremind)
InnoDB is a storage engine for MySQL, but it can also be used
as a standlone file format. This adapter adds a SQL interface
to InnoDB that uses Calcite rather than MySQL.

This adapter handles Sarg by expanding to an OR of ranges. A
better implementation of Sarg is probably possible. The
current implementation can only handle Sargs that result in
an AND, e.g. x >= 10 AND x <= 20). But we ought to handle
Sargs that can result in an OR of ANDs. E.g. the SQL
  x BETWEEN 10 AND 20 OR c > 30
becomes a single RexCall
   SEARCH(x, Sarg([10, 20], (30, +inf)))
and results in an OR of ANDs, '(x >= 10 AND x <= 20) OR (x > 30)'.

Tweaks (Julian Hyde):
* Add Holder.accept
* Make IndexCondition immutable
* Move computation out of InnodbFilter's constructor

Close apache/calcite#1996


Former-commit-id: ebefe52f577da5c744ea863480ed3c98b7ef2e99