[CALCITE-4170] Improve simplification of "<>" predicates
We can now simplify 'a = 1 or a <> 1' to 'true'. To achieve
this, we simplify the second term in a context where we know
'a = 1' is false. This entails a RangeSet inside
simplifyUsingPredicates, so that we can represent <> the same
way that we represent =, > etc. All of the others are a Range,
but <> requires a two-entry RangeSet ('a < 1 or a > 1').


Former-commit-id: ca6c590718599e0b3420757b775994e1075843cc