[CALCITE-4221] Update stale integration tests in Druid adapter
Update doc of DruidAdapterIT and DruidAdapter2IT to reflect the
new dockerized Druid setup.

The changes in the following tests are due to CALCITE-4202:

* testSelectDistinctWiki
* testSelectDistinctWikiNoColumns
* testSelectDistinctWikiNoTables
* testSelectDistinct
* testFilterUnionPlan
* testDistinctCount
* testPushComplexFilter
* testCountOnMetric
* testCountOnMetricRenamed
* testPushEqualsCastDimension
* testPushNotEqualsCastDimension
* testComplexExpressionsIsNull

The changes in the following tests are due to CALCITE-4212:

* testSelectTimestampColumnNoTables1
* testUnionPlan
* testProject
* testTimeExtractThatCannotBePushed
* testSubStringExpressionFilter

The changes in the following tests are due to both CALCITE-4212 and
CALCITE-4202:

* testUnionPlan
* testDistinctCountOnMetric
* testDistinctCountOnMetricRenamed
* testDistinctCountWhenApproxResultsNotAccepted
* testFilterClauseWithMetricRefAndAggregates

The rest of the changes are normal improvements of the optimizer.

Problematic tests are skipped with the appropriate JIRA references
describing the problem.


Former-commit-id: fb8961528ff559e109612465cdef3e62eee852e2