Refactor SqlParserTest
Store StringAndPos rather than the SQL string in a test. This
makes it easier to write positive and negative tests on same
SQL string, because we don't have to keep the carets (that
indicate the error position) in the SQL string. If you want
real carets in a SQL string, you can now write "^^". Remove
the sansCarets() method, as it is no longer necessary.

Pass SqlDialect into negative and expression tests.

Add constants in SqlParserTest for common dialects.


Former-commit-id: 0c5bb1a345f39cab3eaa3b95a77c1e8831e667e1