[CALCITE-4394] When generating code for a function call, take the inferred types of the operands into account
Refactor SqlOperatorBaseTest.testConcatFunc for the various
variants of CONCAT in PostgreSQL, MySQL, Oracle dialects.

Fix code generation that uses built-in functions with
varargs. In particular, SqlFunctions has
concat(String, String) and concat(String...). The fix for
[CALCITE-3563] changed the information used by the
implementor from Method to Class + methodName, which has a
couple of problems: it is lossy if the method is overloaded,
and does not allow us to see whether the method has varargs
parameters. In this commit, we change it back, and add
different code path for varargs methods. We now require
varargs methods to have different names to non-vargs methods.


Former-commit-id: e819b4611e883c54708a75f6856300462c92b8ae