[CALCITE-4408] Implement Oracle SUBSTR function (James Starr)
Add an operand type checker for the Oracle SUBSTR. Becauses
semantics are different to standard SUBSTRING, implement by
translating to SUBSTRING with a CASE expression to handle the
differences.

The implementation assumes that SUBSTRING is consistent with
BigQuery, not ISO Standard SQL, and may need rework when
  [CALCITE-4427] Make SUBSTRING operator comply with ISO
  standard SQL
is fixed.

Rework so that it gives the right result for very negative
start; add tests for MySQL and PostgreSQL variants of SUBSTR
(Julian Hyde).

Close apache/calcite#2271


Former-commit-id: 0327135376246eb58e1703fc13d555efaf389ada