[CALCITE-4514] When merging RelSets, fine-tune which set is merged into which, for efficiency (Botong Huang)
If one set is the parent of the other, merge into the child.

If both sets are parents of the other, or if neither set is
parent of the other, merge into the set that is more popular
(is referenced as child of more relational expressions), or
has more children, or is older. The last condition ensures
determinism.

Close apache/calcite#2358


Former-commit-id: d4e1eeac6e2142d978bfaf216e85aaaf096e9a8d