[CALCITE-4564] Initialization context for non-static user-defined functions (UDFs)
When Interpreter or code generated by RexToLixTranslator calls
a UDF, invoke its constructor only once. There will therefore
be one instance of the UDF for the whole execution, and that
instance can be used to store state. Typical state will be
preparation work (e.g. compiling a regular expression) to make
each invocation faster.

Add 'interface FunctionContext' (marked as experimental). If
a UDF has a public constructor with FunctionContext as a
parameter, this will be called in preference to the default
constructor.

In InterpreterTest, allow more than one query to be executed
per test method.

Close apache/calcite#2395


Former-commit-id: b2e9e6cba1e2ce28368d1281f527a9e53f4628ca