[CALCITE-4620] Join on CASE causes AssertionError in RelToSqlConverter
The fix is to remove the logic that generates ON conditions
and instead use the logic that is used for all other
expressions. This logic is more widely used, and therefore
already handles cases such as CASE. (Previous bugs such as
[CALCITE-3207] LIKE, [CALCITE-4610] Sarg, and
[CALCITE-1422] IS NULL have been caused by this unnecessary
split.)

We have retained the logic that reverses 't1.id = t0.id' to
't0.id = t1.id' if 't0' is from the left side of the join.

We have reverted one test case, which was accidentally
seeing 'mandatoryColumn IS NULL' simplified to 'FALSE',
because the simplification was not the point of the test.


Former-commit-id: af5444ffdfd9e836b052ee58abec9430d55d4139