[CALCITE-3745] CompileException in UnitCompiler when using multiple class loaders
Janino is not able to determine the correct compiler for loading a
given class thus it fails with CompileException.

1. Use the class loader of the calling class instead of the default.
2. Upgrade Janino to 3.1.6 version which contains the fix for
https://github.com/janino-compiler/janino/pull/150 necessary for
passing correctly the class loader to the underlying methods.
3. Add/Update CheckerFramework .astub files for setParentClassLoader
method based on the changes in the Janino APIs. Methods have moved in
the new version so it is necessary to reflect the changes in the stub
files.
4. Enforce class loader is not null before calling
CompilerFactoryFactory#getDefaultCompilerFactory method. The parameter
does not have a nullability annotation so by default CheckerFramework
considers it as not-null. Indeed, if null is given as parameter a NPE
will be raised from within that method.

Close apache/calcite#2449


Former-commit-id: 9c0e3130e6692d1960a34a680dc13d11083ff1c8