Remove obsolete/misleading comments in RelOptUtil#classifyFilters
The claim that filters are not pushed in the NULL generating side is not
true (at least not anymore).

Consider the following query:

SELECT *
FROM emp e
LEFT JOIN dept d
  ON d.deptno > 20

In this case the filter in the ON clause can be pushed in the right
which is the NULL generating side. This is already done by the
FilterJoinRule generating a plan similar to the query below.

SELECT *
FROM emp e
LEFT JOIN (SELECT * FROM dept d WHERE d.deptno > 20) ON true


Former-commit-id: b03ca2fb8fea0b45483697c5796e96898a70646f