Add news item for release 1.28.0
Jacques' description of breaking changes is rather too long
for the release notes, so I added it to the news item, and
referenced it from the release notes.

Make other fixes to the release notes, such as the poorly
escaped '<>' in one commit message.

Add permanent anchors (qualified by release name) to the
release notes, so that #fixes for one release doesn't become
that.

Prepare for next development iteration by changing version
in gradle.properties.

Close apache/calcite#2584


Former-commit-id: c6a9599ceabb6e1f703d3c0018b8daf778ae46f8