[CALCITE-5105] Add MEASURE type and AGGREGATE aggregate function
The MEASURE type is internal. A RexNode expression that
contains a measure and evaluates to an INTEGER will have the
type MEASURE<INTEGER>, using a parameterized SQL type
similar to ARRAY<INTEGER>. But the same measure column, as
seen from SQL, will just have type INTEGER. The
parameterized type helps us keep things straight in the
relational algebra as we apply planner rules.

The AGGREGATE function belongs to the new CALCITE function
library. To use it, add 'lib=calcite' to your connect string.

Add a new validator configuration parameter,
`boolean SqlValidator.Config.nakedMeasures()`. The query

  SELECT deptno, AGGREGATE(avg_sal)
  FROM emp
  GROUP BY deptno

is valid if `avg_sal` is a measure. If `nakedMeasures` is
true, then the following query is a valid shorthand for it:

  SELECT deptno, avg_sal
  FROM emp
  GROUP BY deptno

In the long term, we would like people to feel comfortable
using the latter form. Measures are not necessarily aggregate
functions, but are just expressions whose value depends on
their context (the current GROUP BY key in an aggregate
query, or the current row in a regular query). And we will
generalize measures to analytic expressions, which are not
necessarily just references to measure columns.

But in the short term, setting the `nakedMeasures` flag to
false provides a level of comfort to people (and tools that
generate SQL) who think of measures as aggregate functions,
and think that measures should only be used in `GROUP BY`
queries.

Extend mock catalog with a table that has measure columns.

Add a new Quidem test, measure.iq. It is disabled because we
don't yet have the means to create measure columns in queries
(or views). That is to come in [CALCITE-4496].

Close apache/calcite#2965


Former-commit-id: 406c913b808b3234464d8c81d7352c4040dd281a