Improve debug message in IterativeRuleDriver
Remove PLANNER logging from the message and better formulate at which
point during the rule matching the cost is printed.

The original message is misleading cause it is not the planner that is
printed but the queue. Moreover, IterativeRuleDriver#toString just
prints the class name and hashcode of the instance which is useless.
Class name might be useful but can be obtained by changing the logger
configuration (pattern layout).


Former-commit-id: 96f4a734417d044552a94d8969fd368b57ff9e8b