[CALCITE-4963] Make it easier to implement interface SqlDialectFactory
To create a SqlDialect, you need an instance of
SqlDialect.Context, but the code to create this was until
now locked in class SqlDialectFactoryImpl. This change moves
that code into a public method
SqlDialects.createContext(DatabaseMetaData). It is now
easier to sub-class SqlDialectFactoryImpl or to directly
implement SqlDialectFactory.

Close apache/calcite#2658


Former-commit-id: 8983e7e82ef65c3a72f06305a676cc2998bf72d6