[CALCITE-4401] SqlJoin.toString throws RuntimeException, "No list started"
To resolve the RuntimeException seen in SqlJoin#toString, add
a "SELECT *" wrap over SqlJoin in SqlJoin#toSqlString. This
prevents SqlJoin#unparse method from calling
SqlPrettyWriter#sep with a null frame, and is a less fragile
fix than to mutate internal frame state.

Extend SqlPrettyWriterFixture with a method
`checkTransformedNode(UnaryOperator<SqlNode> transform)`
that does do the same as `check()` but applies a transform
function that maps from the root to another node in the tree.

Close apache/calcite#2757


Former-commit-id: 10a3b1a3f5399851bc31ff665a999411edffbcc8